Fix MD5 checking
